Improving the code structure of ReSolver1d.cc: the array indices are now size_t, consistency improved in layer counting, upper/lower node identifiers, etc. Unsafe unsigned/signed comparisons have been removed, loops are adjusted to deal with unsigned int and confusing variables keeping track of which layer is bowing up are removed, as they are not so important. Calculation results should, however, be exactly the same as before this commit. Otherwise, something went wrong.

+++ snowpack/snowpackCore/ReSolver1d.h
@@ -91,7 +91,7 @@ class ReSolver1d {
 		void SetSoil(SoilTypes type, double *theta_r, double *theta_s, double *alpha, double *m, double *n, double *ksat, double *he);
 
 		// Solvers
-		int TDMASolver (int n, double *a, double *b, double *c, double *v, double *x);
+		int TDMASolver (size_t n, double *a, double *b, double *c, double *v, double *x);
 		int pinv(int m, int n, int lda, double *a);
 };
 #endif //End of WaterTransport.h
